Refactoring result objects into containers.py module
